% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_extract.R
\name{f_FDC_x}
\alias{f_FDC_x}
\title{Cumulative frequency curve at a specific percentile}
\usage{
f_FDC_x(x, probs.FDC = 0.1)
}
\arguments{
\item{x}{a vector of real, value of the time series.}

\item{probs.FDC}{a real, between 0 and 1. Specific probability of occurence of the Cumulative
frequency curve (default: "0.1).}
}
\value{
a real, value corresponding to the Cumulative frequency curve at the probalbility probs.FDC
}
\description{
Function to compute the Cumulative frequency Curve of a time serie at
a specific occurrence probability. In Hydrology, the cumulative frequency curve
of a Discharge time serie is called "Flow Duration curve".
}
\examples{
f_FDC_x(x=vect.Q)
f_FDC_x(x=vect.Q,probs.FDC=0.5)
read.timeSeries(datapath=datapath,ind.catch=c(1,9,14,5,80))
}
