% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests_function.R
\name{FDC_lowvol}
\alias{FDC_lowvol}
\title{Volume deficit}
\usage{
FDC_lowvol(x)
}
\arguments{
\item{x}{a vector of real, value of the time series.}
}
\value{
a real, number corresponding to the value of the FDC at the
probability probs.FDC
}
\description{
Function to compute the variable "Volume deficit": sum of
differences between the values of the time series and the cumulative
frequency curve (probability of occurence against value) at the percentile
0.15.
}
\examples{
vect.Q= exp(seq(from = 0.01,to = 2,by = 0.01))
FDC_lowvol(x=vect.Q)
}
