% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{readFiles.TS}
\alias{readFiles.TS}
\title{Load and process data}
\usage{
readFiles.TS(
  datapath,
  ind.catch = NULL,
  NA.code = "-99.000",
  col.date = c(1, 2, 3)
)
}
\arguments{
\item{datapath}{vector of character, path of the files to load.}

\item{ind.catch}{vector of integer, index of station to keep (default: NULL, all files are considered).}

\item{NA.code}{vector of strings, codes of the NA values in the files to be read (default: "-99.000").}

\item{col.date}{vector of integer, column number for year, month, day, if length == 1,
it is already a date format.}
}
\value{
a list of two objects (data, a unique data.tibble containing all
the data grouped by file; info, a data.tibble with the file and matching groups)
}
\description{
Read data from specified files and orgnaize it in the proper format for variable extraction.
}
\examples{
read.timeSeries(datapath=datapath)
read.timeSeries(datapath=datapath,ind.catch=1:51)
read.timeSeries(datapath=datapath,ind.catch=c(1,9,14,5,80))
read.timeSeries(datapath=datapath,,NA.code="-9999")
read.timeSeries(datapath=datapath,,NA.code=c("-9999","-1"))
}
