% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests_function.R
\name{regionalTrend}
\alias{regionalTrend}
\title{Regional consistency test}
\usage{
regionalTrend(X, level = 0.1)
}
\arguments{
\item{X}{numeric matrix, n*p, n=number of time steps, p=number of sites}

\item{level}{numeric in (0,1), level of the test}
}
\value{
A list with the following fields:
\enumerate{
\item H: logical, reject (true) or do not reject (false) H0
\item P: p-value of the test
\item STAT: test statistics
\item TREND: estimate of the common trend
\item DATA: Normal-scored data used in the test
}
}
\description{
Test for the existence of a common trend in several series
}
\details{
Handling of missing data: incomplete years (i.e. with a missing value for at least one site)
are entirely discarded. It might hence be beneficial to remove patchy sites from X.
}
\examples{
set.seed(123456) # make example reproducible
data(nhtemp) # 	Average Yearly Temperatures in New Haven
# Create 15 noisy versions of the time series
X=matrix(rep(nhtemp,15)+rnorm(length(nhtemp)*15),nrow=length(nhtemp))
# put a few missing values
X[10:15,1]=NA;X[20,5]=NA
# Apply regional test
res=regionalTrend(X)
# Plot normal-scored data and common trend
matplot(res$DATA$time,res$DATA[,3:NCOL(res$DATA)])
lines(res$DATA$time,res$TREND*res$DATA$time)
}
\references{
Renard, B., et al. (2008), Regional methods for trend detection: Assessing field significance
and regional consistency, Water Resour. Res., doi:10.1029/2007WR006268
}
