% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationQ_3catch.R
\docType{data}
\name{StationQ_3catch}
\alias{StationQ_3catch}
\title{Few Discharge time series from the French national discharge database "Banque HYDRO"}
\format{
A list of two tibble objects:
\describe{
\item{data}{Date to analyse: different classifications/groups and corresponding value}
\item{info}{Info corresponding to each group}
}
}
\usage{
StationQ_3catch
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\keyword{datasets}
