% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationQ_TS.R
\docType{data}
\name{StationQ_TS}
\alias{StationQ_TS}
\title{Few Discharge time series from the French national discharge database "Banque HYDRO"}
\format{
A list of two tibble objects:
\describe{
\item{data}{Date to analyse: different classifications/groups and corresponding value}
\item{info}{Info corresponding to each group}
}
}
\source{
\url{https://www.hydroportail.developpement-durable.gouv.fr/}
}
\usage{
StationQ_TS
}
\description{
The streamflow station were selected by Giuntoli et al. (2013)
Pour les donnees elles-meme, si tu publies un papier, le mieux c'est de faire un depot Zenodo
de la version des donnees que tu as utilisee, et de rajouter une phrase du genre 'the latest
version of the time series can be downloaded from the HydroPortail
(https://www.hydroportail.developpement-durable.gouv.fr/) managed by the French Ministry of Ecology'.
Note que le site n'est pas encore ouvert au grand public, mais ce sera le cas debut 2022.
}
\keyword{datasets}
